Refactor code -- PagesVersionHandler
Delete com.ibm.ws.javaee.version PagesVersion.java

Review comments
