--- dev/com.ibm.ws.javaee.version/src/com/ibm/ws/javaee/version/PagesVersion.java
@@ -1,18 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021,2022 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package com.ibm.ws.javaee.version;
-
-// TODO: This would be nice to remove.  To do that,
-//       service references would need to be removed.
-
-public class PagesVersion {
-    // EMPTY
-}
