+++ dev/com.ibm.ws.jsp.2.3/bnd.bnd
@@ -24,7 +24,7 @@ com.ibm.ws.jsp23.translator.visitor.validator.ElValidatorExtFactoryImpl, \
 com.ibm.ws.jsp23.translator.visitor.generator.GeneratorUtilsExtFactoryImpl
 
 Export-Package: \
-   com.ibm.ws.jsp.speclevel, \
+	com.ibm.ws.jsp.speclevel, \
 	com.ibm.ws.jsp23.webcontainerext;version="2.3", \
 	com.ibm.ws.jsp23.webcontainerext.ws;version="2.3", \
 	com.ibm.ws.jsp23.translator.visitor.validator;version="2.3", \
