+++ dev/com.ibm.ws.jsp/bnd.bnd
@@ -159,6 +159,5 @@ instrument.classesExcludes: com/ibm/ws/jsp/resources/*.class
 	com.ibm.ws.artifact.url;version=latest, \
 	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
 	com.ibm.ws.org.apache.taglibs.standard;version=latest, \
-	com.ibm.ws.javaee.version;version=latest, \
 	io.openliberty.jakarta.servlet.jsp.jstl.2.0.internal;version=latest, \
 	io.openliberty.el.internal.cdi;version=latest
